start move to new directory
