+++ src/parser/mage/frost/CombatLogParser.js
@@ -12,7 +12,7 @@ import BrainFreeze from './modules/features/BrainFreeze';
 import IceLance from './modules/features/IceLance';
 import ThermalVoid from './modules/features/ThermalVoid';
 import GlacialSpike from './modules/features/GlacialSpike';
-import BoneChilling from './modules/features/BoneChilling';
+import BoneChilling from './modules/talents/BoneChilling';
 import RuneOfPower from '../shared/modules/features/RuneOfPower';
 import MirrorImage from '../shared/modules/features/MirrorImage';
 import ArcaneIntellect from '../shared/modules/features/ArcaneIntellect';
@@ -28,7 +28,6 @@ class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
     checklist: Checklist,
 
-
     // Features
     abilities: Abilities,
     alwaysBeCasting: AlwaysBeCasting,
@@ -37,21 +36,33 @@ class CombatLogParser extends CoreCombatLogParser {
 	  wintersChill: WintersChill,
 	  brainFreeze: BrainFreeze,
     iceLance: IceLance,
-	  thermalVoid: ThermalVoid,
-	  glacialSpike: GlacialSpike,
     damageDone: [DamageDone, { showStatistic: true }],
-    runeOfPower: RuneOfPower,
-    mirrorImage: MirrorImage,
     arcaneIntellect: ArcaneIntellect,
-    splittingIce: SplittingIce,
-    boneChilling: BoneChilling,
     waterElemental: WaterElemental,
 
-    //Traits
+    // region Talents (T30 and T75 don't need analyzers)
+    // T15 TODO - Ice Nova, Lonely Winter
+    boneChilling: BoneChilling,
+
+    // T45 TODO - Incanters Flow
+    mirrorImage: MirrorImage,
+    runeOfPower: RuneOfPower,
+
+    // T60 TODO - Frozen Touch, Chain Reaction, Ebonbolt
+
+    // T90 TODO - Freezing Rain, Comet Storm
+    splittingIce: SplittingIce,
+
+    // T100 TODO - Ray of Frost
+    thermalVoid: ThermalVoid,
+    glacialSpike: GlacialSpike,
+    // endregion
+
+    // Traits
     wintersReach: WintersReach,
     whiteout: Whiteout,
 
-	  //Cooldowns
+	  // Cooldowns
     frozenOrb: FrozenOrb,
     coldSnap: ColdSnap,
